+++ src/helpers/Cp.php
@@ -1189,6 +1189,161 @@ JS, [
     }
 
     /**
+     * Returns address fields’ HTML (sans country) for a given address.
+     *
+     * @param Address $address
+     * @return string
+     * @since 4.0.0
+     */
+    public static function addressFieldsHtml(Address $address): string
+    {
+        $formatRepo = Craft::$app->getAddresses()->getAddressFormatRepository()->get($address->countryCode);
+
+        $requiredFields = array_flip($formatRepo->getRequiredFields());
+        $visibleFields = array_flip(array_merge(
+                $formatRepo->getUsedFields(),
+                $formatRepo->getUsedSubdivisionFields(),
+            )) + $requiredFields;
+
+        return
+            static::textFieldHtml([
+                'label' => $address->getAttributeLabel('addressLine1'),
+                'id' => 'addressLine1',
+                'name' => 'addressLine1',
+                'value' => $address->addressLine1,
+                'required' => isset($requiredFields['addressLine1']),
+                'errors' => $address->getErrors('addressLine1'),
+            ]) .
+            static::textFieldHtml([
+                'label' => $address->getAttributeLabel('addressLine2'),
+                'id' => 'addressLine2',
+                'name' => 'addressLine2',
+                'value' => $address->addressLine2,
+                'required' => isset($requiredFields['addressLine2']),
+                'errors' => $address->getErrors('addressLine2'),
+            ]) .
+            self::_subdivisionField(
+                $address,
+                'administrativeArea',
+                isset($visibleFields['administrativeArea']),
+                isset($requiredFields['administrativeArea']),
+                [$address->countryCode],
+                true,
+            ) .
+            self::_subdivisionField(
+                $address,
+                'locality',
+                isset($visibleFields['locality']),
+                isset($requiredFields['locality']),
+                [$address->countryCode, $address->administrativeArea],
+                true,
+            ) .
+            self::_subdivisionField(
+                $address,
+                'dependentLocality',
+                isset($visibleFields['dependentLocality']),
+                isset($requiredFields['dependentLocality']),
+                [$address->countryCode, $address->administrativeArea, $address->locality],
+                false,
+            ) .
+            Html::beginTag('div', ['class' => 'flex-fields']) .
+            static::textFieldHtml([
+                'fieldClass' => array_filter([
+                    'width-50',
+                    !isset($visibleFields['postalCode']) ? 'hidden' : null,
+                ]),
+                'label' => $address->getAttributeLabel('postalCode'),
+                'id' => 'postalCode',
+                'name' => 'postalCode',
+                'value' => $address->postalCode,
+                'required' => isset($requiredFields['postalCode']),
+                'errors' => $address->getErrors('postalCode'),
+            ]) .
+            static::textFieldHtml([
+                'fieldClass' => array_filter([
+                    'width-50',
+                    !isset($visibleFields['sortingCode']) ? 'hidden' : null,
+                ]),
+                'label' => $address->getAttributeLabel('sortingCode'),
+                'id' => 'sortingCode',
+                'name' => 'sortingCode',
+                'value' => $address->sortingCode,
+                'required' => isset($requiredFields['sortingCode']),
+                'errors' => $address->getErrors('sortingCode'),
+            ]) .
+            Html::endTag('div'); // .flex-fields
+    }
+
+    private static function _subdivisionField(
+        Address $address,
+        string $name,
+        bool $visible,
+        bool $required,
+        ?array $parents,
+        bool $spinner,
+    ): string {
+        $value = $address->$name;
+        $options = Craft::$app->getAddresses()->getSubdivisionRepository()->getList($parents, Craft::$app->language);
+
+        if ($options) {
+            // Persist invalid values in the UI
+            if ($value && !isset($options[$value])) {
+                $options[$value] = $value;
+            }
+
+            if ($spinner) {
+                $errors = $address->getErrors($name);
+                $input =
+                    Html::beginTag('div', [
+                        'class' => ['flex', 'flex-nowrap'],
+                    ]) .
+                    static::selectizeHtml([
+                        'id' => $name,
+                        'name' => $name,
+                        'value' => $value,
+                        'options' => $options,
+                        'errors' => $errors,
+                    ]) .
+                    Html::tag('div', '', [
+                        'id' => "$name-spinner",
+                        'class' => ['spinner', 'hidden'],
+                    ]) .
+                    Html::endTag('div');
+
+                return static::fieldHtml($input, [
+                    'fieldClass' => !$visible ? 'hidden' : null,
+                    'label' => $address->getAttributeLabel($name),
+                    'id' => $name,
+                    'required' => $required,
+                    'errors' => $errors,
+                ]);
+            }
+
+            return static::selectizeFieldHtml([
+                'fieldClass' => !$visible ? 'hidden' : null,
+                'label' => $address->getAttributeLabel($name),
+                'id' => $name,
+                'name' => $name,
+                'value' => $value,
+                'options' => $options,
+                'required' => $required,
+                'errors' => $address->getErrors($name),
+            ]);
+        }
+
+        // No preconfigured subdivisions for the given parents, so just output a text input
+        return static::textFieldHtml([
+            'fieldClass' => !$visible ? 'hidden' : null,
+            'label' => $address->getAttributeLabel($name),
+            'id' => $name,
+            'name' => $name,
+            'value' => $value,
+            'required' => $required,
+            'errors' => $address->getErrors($name),
+        ]);
+    }
+
+    /**
      * Renders a field layout designer.
      *
      * @param FieldLayout $fieldLayout